backend/requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis pour cache
redis==5.0.1

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Collecteurs
feedparser==6.0.11
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2

# HTML parsing
lxml==5.1.0
html5lib==1.1

# Utilitaires
pytz==2023.3

# Logging et monitoring
python-json-logger==2.0.7

# Testing (optionnel)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0